Move RuntimeArgs classes to correct package Use static for callback
