+++ modules/runtime-args/src/main/java/com/gluonhq/attach/runtimeargs/RuntimeArgsService.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.attach.runtime;
+package com.gluonhq.attach.runtimeargs;
 
 import com.gluonhq.attach.util.Services;
