+++ modules/runtime-args/src/main/java/com/gluonhq/attach/runtimeargs/impl/DesktopRuntimeArgsService.java
@@ -25,9 +25,11 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package com.gluonhq.attach.runtimeargs.impl;
 
 /**
- * Primary API package for Attach - RuntimeArgs plugin,
- * contains the interface {@link com.gluonhq.attach.runtime.RuntimeArgsService} and related classes.
+ * An implementation of RuntimeArgsService on desktop
  */
-package com.gluonhq.attach.runtime;
+public class DesktopRuntimeArgsService extends DefaultRuntimeArgsService {
+
+}
