+++ modules/runtime-args/src/main/java/com/gluonhq/attach/runtimeargs/impl/DummyRuntimeArgsService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2019, Gluon
+ * Copyright (c) 2018, Gluon
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,36 +25,10 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.attach.runtime.impl;
+package com.gluonhq.attach.runtimeargs.impl;
 
-import com.gluonhq.attach.runtime.RuntimeArgsService;
-import com.gluonhq.attach.util.Constants;
+import com.gluonhq.attach.runtimeargs.RuntimeArgsService;
 
-/**
- * An implementation of the
- * {@link RuntimeArgsService RuntimeArgsService} for the
- * iOS platform. 
- */
-public class IOSRuntimeArgsService extends DefaultRuntimeArgsService {
-
-    static {
-        System.loadLibrary("RuntimeArgs");
-        initRuntimeArgs();
-    }
-
-    public IOSRuntimeArgsService() {
-        if ("true".equals(System.getProperty(Constants.ATTACH_DEBUG))) {
-            enableDebug();
-        }
-    }
-
-    // Native
-    
-    private static native void initRuntimeArgs();
-    private static native void enableDebug();
-
-    // callback
-    private void processRuntimeArgs(String key, String value) {
-        super.fire(key, value);
-    }
+// no-op
+public abstract class DummyRuntimeArgsService implements RuntimeArgsService {
 }
